feat(tests): Playwright E2E POC — chassis validation against real Kestrel + Chromium (#16 slice b)
Lands the third leg of #16 — Playwright drives a real Chromium against
real Kestrel-hosted BookTracker.Web on a random port, with the existing
SqlServerContainer providing the DB.

Two tests in ChassisSmokeTests:
  - HomePage_Renders — server bound, browser fetched, navbar contains
    the brand. If this fails, every other E2E test is doomed.
  - SeriesListPage_RendersHeading — one-hop navigation off the home
    page; catches routing + page-level rendering regressions.

POC scope on purpose: deliberately *not* the form-submit + redirect
flow that motivated this work in #168. Blazor's static-SSR-with-FormName
forms vs InteractiveServer interactive submit is a real
investigation that warrants its own follow-up — Playwright fills the
input via DOM events but the form post then takes the static-SSR
path, where VM.Input isn't a [SupplyParameterFromForm] parameter, so
save no-ops. Documented in the test class summary.

Implementation notes:
- ProgramSetup extracted from Program.cs's top-level statements so the
  test fixture can build a configured WebApplication without
  duplicating DI / middleware wiring. Program.Main becomes a thin
  wrapper: Build → RunMigrations → RunAsync.
- WebApplicationFactory<Program> rejected because its Server property
  hardcodes a cast to TestServer; calling ProgramSetup.Build directly
  bypasses the helper while going through the same pipeline.
- MapStaticAssets gracefully skips when the manifest isn't present
  (test host runs as BookTracker.Tests, not BookTracker.Web). Razor
  pages render fine without the static-asset endpoint mapping; the
  JS / CSS that's normally cache-busted just 404s, which Playwright
  doesn't care about for these assertions.
- Listen on 127.0.0.1:0 (NOT ListenLocalhost(0) — that throws on
  dynamic-port binding to the dual-stack listener).

Local one-time setup after build:
  pwsh BookTracker.Tests/bin/Debug/net10.0/playwright.ps1 install chromium

CI integration deferred. PR job filter unchanged
(Unit|Component|Integration) — E2E gets its own job in a follow-up.

Suite: 342/1/0 with E2E running (was 340/1/0 before).
